Move find_file to `utils.rb`
And delete search_utils - its reasons for existence are no longer
applicable now the module has shrunk so much.
