+++ lib/inventoryware/utils/utils.rb
@@ -60,5 +60,22 @@ Please create it before continuing"
       end
       return true
     end
+
+    # return a single file from glob, print error if >/< than 1 found
+    def self.find_file(search_val, dir)
+      results = Dir.glob(File.join(dir, "#{search_val}*"))
+        if results.empty?
+          puts "No files found for '#{search_val}' in #{File.expand_path(dir)}"
+        elsif results.length > 1
+          file_names = results.map { |p| File.basename(p, File.extname(p)) }
+          # if the results include just the search val, return that path
+          if file_names.include?(search_val)
+            return results.select { |p| p[/#{search_val}\..*$/] }
+          end
+          puts "Ambiguous search term '#{search_val}' - possible results are:"
+          file_names.each_slice(3).each { |p| puts p.join("  ") }
+        end
+      return results
+    end
   end
 end
